+++ cmd/main.go
@@ -249,6 +249,14 @@ func main() {
 		setupLog.Error(err, "Failed to create controller", "controller", "omnikubeconfigexport")
 		os.Exit(1)
 	}
+	if err := (&controller.OmniHelmReleaseReconciler{
+		Client:       mgr.GetClient(),
+		Scheme:       mgr.GetScheme(),
+		SecretReader: mgr.GetAPIReader(),
+	}).SetupWithManager(mgr); err != nil {
+		setupLog.Error(err, "Failed to create controller", "controller", "omnihelmrelease")
+		os.Exit(1)
+	}
 	if err := (&controller.OmniCiliumReconciler{
 		Client: mgr.GetClient(),
 		Scheme: mgr.GetScheme(),
@@ -266,6 +274,7 @@ func main() {
 			{name: "OmniCluster", setup: webhookv1alpha1.SetupOmniClusterWebhookWithManager},
 			{name: "OmniConnection", setup: webhookv1alpha1.SetupOmniConnectionWebhookWithManager},
 			{name: "OmniControlPlane", setup: webhookv1alpha1.SetupOmniControlPlaneWebhookWithManager},
+			{name: "OmniHelmRelease", setup: webhookv1alpha1.SetupOmniHelmReleaseWebhookWithManager},
 			{name: "OmniKubeconfigExport", setup: webhookv1alpha1.SetupOmniKubeconfigExportWebhookWithManager},
 			{name: "OmniMachine", setup: webhookv1alpha1.SetupOmniMachineWebhookWithManager},
 			{name: "OmniWorkers", setup: webhookv1alpha1.SetupOmniWorkersWebhookWithManager},