+++ internal/controller/child_watch.go
@@ -91,6 +91,22 @@ func addonRequestsForCluster(ctx context.Context, c client.Client, object client
 	return sortRequests(requests)
 }
 
+func helmReleaseRequestsForCluster(ctx context.Context, c client.Client, object client.Object) []reconcile.Request {
+	releaseList := &omniv1alpha1.OmniHelmReleaseList{}
+	if err := c.List(ctx, releaseList, client.InNamespace(object.GetNamespace())); err != nil {
+		return nil
+	}
+
+	var requests []reconcile.Request
+	for _, item := range releaseList.Items {
+		if item.Spec.ClusterRef.Name == object.GetName() {
+			requests = append(requests, reconcile.Request{NamespacedName: client.ObjectKey{Namespace: item.Namespace, Name: item.Name}})
+		}
+	}
+
+	return sortRequests(requests)
+}
+
 func ciliumRequestsForCluster(ctx context.Context, c client.Client, object client.Object) []reconcile.Request {
 	ciliumList := &omniv1alpha1.OmniCiliumList{}
 	if err := c.List(ctx, ciliumList, client.InNamespace(object.GetNamespace())); err != nil {